Fix random generation in tests
My original version used permuted. This does not handle symmetries
with sector_ind_len > 1 correctly, as the different rows of a
sector_array are shuffeled *independently* by `permuted`.

Ludwig fixed this by using `shuffle` instead.
This caused a proplem with in-place mutation of caches.

This is now fixed by introducing a custom `shuffled` function,
which `shuffle`s a copy.