Update ElementarySpace class
Change to a new mental model for ElementarySpace.
We now take seriously the idea of an explicit duality flag,
which now means "consider the other attrs, then take the dual of that space".

Main change is a new attribute `defining_sectors` of ElementarySpace
which is distinct from the `sector_decomposition`.

Introduces a few bugs in abelian and fusiontree backends.
To be fixed in future commits.